fix: deduplicate helpers
