+++ lib/core/utils/network_error_utils.dart
@@ -1,0 +1,16 @@
+import 'dart:async';
+import 'dart:io';
+
+bool isNetworkError(dynamic error) {
+  if (error is SocketException || error is TimeoutException) {
+    return true;
+  }
+
+  final errorString = error.toString().toLowerCase();
+  return errorString.contains('network') ||
+      errorString.contains('connection') ||
+      errorString.contains('hostname') ||
+      errorString.contains('socket') ||
+      errorString.contains('timeout') ||
+      errorString.contains('no address');
+}
